backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0

# AI & ML
google-generativeai==0.4.1
assemblyai==0.22.0
scikit-learn
scipy
numpy

# Database & External Services
supabase==2.3.0
spotipy==2.23.0
